Renamed and reorganized components
Highlights:
 * The main component at /record-search is now called RecordSearch. What was called RecordSearch is now called SearchPanel
 * Components for the Search feature have a nested file structure
 * All Type interface definitions for the data models are called [Thing]Data instead of [Thing]Type
 * Type interface definitions are located in the relevant component folders and can get imported to the redux code when necessary
 * renamed clearSearchRecord to clearRecord
 * Deleted the inconsistent use of "container" components
