+++ src/frontend/src/components/RecordSearch/Record/RecordSummary/CasesSummary.tsx
