+++ src/frontend/src/components/RecordSearch/Record/RecordSummary/ChargesList.tsx
