+++ src/frontend/src/components/RecordSearch/Record/index.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
-import Cases from '../Cases';
-import RecordSummary from '../RecordSummary';
-import { Record } from './types';
+import Cases from './Cases';
+import RecordSummary from './RecordSummary';
+import { RecordData } from './types';
 
 interface Props {
-  record: Record;
+  record: RecordData;
 }
 
-export default class SearchResults extends React.Component<Props> {
+export default class Record extends React.Component<Props> {
   render() {
     const errors = ( this.props.record.errors ?
       this.props.record.errors.map(((errorMessage: string, errorIndex: number) => {
         const id= "record_error_" + errorIndex;
-        
-        const errorMessageArray = errorMessage.split(/(\[.*?\])/g); 
+
+        const errorMessageArray = errorMessage.split(/(\[.*?\])/g);
         const errorMessageHTML = errorMessageArray.map(function (element) {
           if (element.match(/^\[.*\]$/)) {
               const caseNumber = element.slice(1, -1);
