--- src/frontend/src/containers/AllRecords.tsx
@@ -1,68 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { AppState } from '../redux/store';
-import { Record } from '../redux/record/types';
-import {
-  searchRecord,
-  clearSearchRecord
-} from '../redux/record/actions';
-import RecordSearch from '../components/RecordSearch';
-import SearchResults from '../components/SearchResults';
-import AllStatus from './AllStatus';
-import { checkOeciRedirect } from '../service/cookie-service';
-
-type Props = {
-  fetchRecord: Function;
-  clearRecord: Function;
-  record?: Record;
-};
-
-class AllRecords extends Component<Props> {
-  componentDidMount() {
-    checkOeciRedirect();
-  }
-
-  componentWillUnmount() {
-    this.props.clearRecord();
-  }
-
-  render() {
-    return (
-      <main className="mw8 center ph2">
-        <RecordSearch fetchRecord={this.props.fetchRecord} />
-        {this.props.record &&
-        ((this.props.record.cases &&
-        this.props.record.cases.length > 0) || this.props.record.errors) ? (
-          <SearchResults record={this.props.record} />
-        ) : (
-          <AllStatus />
-        )}
-        <div className="bg-white shadow mv4 pa4 br3">
-          <h2 className="fw6 mb3">Assumptions</h2>
-          <p className="mb3">We are only able to access your public Oregon records.</p>
-          <p className="mb2">Your analysis may be different if you have had cases which were:</p>
-          <ul className="lh-copy pl4 mw6">
-            <li className="mb2">Previously expunged</li>
-            <li className="mb2">From States besides Oregon within the last ten years</li>
-            <li className="mb2">From Federal Court within the last ten years</li>
-            <li className="mb2">From local District Courts, e.g. Medford Municipal Court (not Jackson County Circuit Court) from within the last ten years</li>
-          </ul>
-        </div>
-      </main>
-    );
-  }
-}
-
-const mapStateToProps = (state: AppState) => {
-  return {
-    record: state.record.record
-  };
-};
-
-export default connect(
-  mapStateToProps,
-  {
-    fetchRecord: searchRecord,
-    clearRecord: clearSearchRecord
-  }
-)(AllRecords);
