+++ src/frontend/src/redux/record/types.ts
@@ -1,35 +1,28 @@
-import {RecordSummaryType} from '../../components/SearchResults/types'
+import {RecordSummaryData, RecordData} from '../../components/RecordSearch/Record/types'
 
 export interface RecordWrapper {
-  record: Record;
+  record: RecordData;
 }
 export interface SearchResponse {
   data: RecordWrapper;
 }
 
-export interface Record {
-  total_balance_Due?: number;
-  cases?: any[];
-  errors?: string[];
-  summary?: RecordSummaryType;
-}
-
 // These constants are used as the 'type' field in Redux actions.
 export const SEARCH_RECORD = 'LOAD_SEARCH_RECORD';
 export const SEARCH_RECORD_LOADING = 'LOAD_SEARCH_RECORD_LOADING';
-export const CLEAR_SEARCH_RECORD = 'CLEAR_SEARCH_RECORD';
+export const CLEAR_RECORD = 'CLEAR_RECORD';
 
 export interface SearchRecordState {
   loading: boolean;
-  record?: Record;
+  record?: RecordData;
 }
 
 interface SearchRecordAction {
   type:
     | typeof SEARCH_RECORD
     | typeof SEARCH_RECORD_LOADING
-    | typeof CLEAR_SEARCH_RECORD;
-  search_record: Record;
+    | typeof CLEAR_RECORD;
+  search_record: RecordData;
 }
 
 // Add other Action types here like so:
